requirements.txt
openpyxl
pandas
python-dateutil
python-dotenv
requests
scikit-learn

lxml
rdflib
SPARQLWrapper
wikidata
wiki-dump-reader
wikipedia-api

neo4j
langchain
langchain-community
langchain-experimental
langchain-neo4j
langchain-openai
openai